view-model: Use GSortListModel internally
This is a required step in order to use GtkTreeListModel later.

GtkTreeListModel doesn't work well when its internal models get
resorted: it assumes the items are gone, not reordered, causing
the row objects to get destroyed and the view to lose track of
things.

This might not be as efficient, but we've got no other option
other than implementing a custom tree list model...