list-base: Ignore double-click while holding Shift or Ctrl
Double clicking an item while holding down Shift can result in opening
a bunch of files in a go, because it selects an interval right before
triggering activation.

This is obviously undersirable, especially if people expect the that
holding down the modifier would actually modify the double click action
to open in a new window or tab (the old canvas view used to do just
that, but that behaviro was inconsistent with the list view anyway).

The safe thing to do here is nothing. Just ignoring the double-click
if Shift or Ctrl are pressed.